requirements-dev.txt
# Development dependencies for data_replication project

# Core dependencies
databricks-connect==17.1.*
databricks-sdk>=0.8.0

# Configuration and validation
pydantic>=2.0.0
pyyaml>=6.0

# Retry mechanism
tenacity>=8.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
coverage>=7.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Development tools
setuptools
wheel
